CellPartinioner.GroupRelatedCells builds an undirected graph with entity and association sets as vertices, and edges corresponding to the mapping fragments. For the association sets in CSpace it also adds edges between the entity sets corresponding to the ends of the association. Edges are also added between the sets corresponding to the principal and dependent tables that are part of foreign key constraints in SSpace. Using a "connected components" algorithm the method creates groups of related cells.
The model defines an association set in CSpace and maps it to an entity set from SSpace (store:Type="Views"). An association set is not defined in SSpace and thus there are no foreign key constraints. This results in a graph with two connected components instead of one, and thus two cell groups, while all the cells are related (in terms of view generation).

The fix checks if the association is many to many, in which case adds edges between the association set and the entity sets corresponding to the association ends.