Updating the build infrastructure to match EFDesigner commit 83817dc932c65180c554c0c9608e69717c1c72d6 - the new build infrstaructure which supports servicing better.
Incorporating changes to the E2ETests.

Updating EFToolsString to be tokenized and the WixId to be automatically
calculated.

Make sure VsIdeTestHostAdapter uses correct versions and updates
EFTools.msbuild to correct ReRunTests target.

This version still pulls in a fixed version of the runtime to the designer
build - that is still to be fixed.