The Annotated Fluent Reference Manual (HasAnnotation APIs for item 57)
This change adds Code First fluent API to allow serializable annotations to be placed on store columns. This is required for flowing information through Code First such as is needed by Zumo and to implement the index attribute. The equivalent APIs for conventions are not included here and will be added in a future check in.

The basic idea is that anytime you can do HasColumnName or MapKey, then you can also do HasAnnotation or HasKeyAnnotation. HasKeyAnnotation requires that the name of the key be provided as well since this seems to be the best way to fit it into the current fluent API classes.