You keep using that initializer. I do not think it means what you think it means. (1709: Initializers and Migrations)
In EF5 a database initializer would completely ignore any Migrations configurations and create a database using defaults which could result in a different database than would be created by Migrations. In EF6 we changed this to throw since if the app is using Migrations then Migrations should be used to manage the database rather than the older initializers. We got feedback on this that people didn't like it because they wanted to have Migrations configured but still use initializers in development/test environments where dropping and re-creating the database is okay. This change causes the older initializers to use a DbMigrationsConfiguration and its associated migrations if discovered so that the older initializers can still be used in the dev/test environment.

Some things to note:
1. The initialer Seed method will be run but and the MigrationsConfiguration Seed method will not be run. This is consistent with the EF5 behavior and seems to match most closely to what people want from this combination. It would also be non-trivial to get the Migrations Seed method to be called without potentially corrupting context state.

2. The initializer will throw an exception if calling DbMigrator.Update would have thrown an exception. Specifically, if auto-migrations are disabled and there are model changes for which no Migration has been added yet.