2008 strikes back... (1832: Structural annotations not working after changeset 00a41e1493)
The code for reading in custom annotations can handle XML attributes by using the attribute value string. However, it can also handle XML elements in some cases in which case the LINQ to XML XElement is placed in the EDM property. This was broken because we were attempting to cast the element to a string which results in null rather than an exception, and hence the EDM property value was null. The fix is to handle XElements as they were before without the cast or any serialization support.