geNeRatE model: Fixing a NullReferenceException happening for Database First if the model file already exists
We would throw a NRE if the user provided a name for model that would conflict with an exsiting file and selected "Generate from Database" option in the wizard (note that the exception was swallowed by VS so it would not bubble to the user). The underlying issue was that to prevent flickering we removed all but the first page from the wizard which made the wizard think we are on the last page and attempt to generate the model. There was also another NRE hidden by the original one - when clicking on the Next button in the wizard the wizard would schedule activating the next page before calling OnDeactivate() for the current page. However the next page (WizardPageDbConfig) does initalization during activation and expects the name of the model to be set which was not set because it was invalid. The fix is to change the logic that prevents flickering - we no longer remove any pages from the wizard and prevent from activating both next and current pages. This prevents flickering and fixes NREs. In addition we used to call Deactivate for a page after the wizard form was closed. This felt wrong since the wizard form owns pages so I changed the order to first call Deactivate and then close the Wizard form.