Little funcy unicorn factories... (1131: Make IDbContextFactory<T> a service)
This changes provides a service and DbConfiguration sugar to allow <see cref="DbContextInfo"/> to create instances of a context that does not have a public, parameterless constructor. This is typically needed to allow design-time tools like Migrations or scaffolding code to use contexts that do not have public, parameterless constructors. Previously this could only be done by including an IDbContextFactory<T> implementation in the same assembly as the context.