Writing on the table... (Fluent API for serializable custom annotations on store types)
This change introduces HasAnnotation for types. This works basically the same way has HasAnnotation for properties. It can be used on the base entity configuration in which case any table that all or part of the entity maps to will get the annotations. It can also be used inside Map calls to put annotations onto different tables for things like TPC, splitting, and many-to-many join tables. In these cases only the table being configured by the given Map call with get the annotations. In cases where multiple configurations map to the same table (e.g. TPH, splitting) the annotations can be specified on one or more of these configurations, but an exception is thrown if two configurations try to set the same annotation to different values. This is the same thing with do on properties.