ColumnMapBabelFish: Fix for "Enum columns bails out with many Includes" (https://entityframework.codeplex.com/workitem/1686)
Whe translating the query tree we replace occurences of enum types with its underlying type. After this is done all typed items we create in later phases of the translation are of primitive types. However the tree contains a column map which is used to generate code that is materializing entities. If this column map had to be rewritten we would change the type of enum properties to its underlying type beacuse the newly created replacement property/variable would be of a primitive type. This resulted in emitting materialization code that was meant to be used with prmitive types but not with enums. This code actually worked fine in majority of cases for non-key properties however in case of key properties we do some additional validation that checks whether the type of the materialized value is the same as the type of the key property and this validation failed because types were different. The fix is not to retain the type of the column if it was of an enum type when rewriting column maps.

which resulted in emitting code that was missing a cast