SnowdenizingProviderServices - making the internal SetParameterValue method public (fixes http://entityframework.codeplex.com/workitem/1867)
In DbProviderServices we follow a pattern where we have public instance methods calling into their protected counterparts to enable creating wrapping providers. We didn't follow this pattern for the SetParameterValue method (it was internal) which caused issues if the provider had some additional logic to process parameter values before setting them - for instance a provider wrapping SqlServerProviderServices was not usable with spatial types where the conversion from EF specific spatial types to Sql Server specific spatial types for parameters happens in the SetDbParameterValue method which was not accessible to the wrapping provider.