IHateRepeatingMyselfIHateRepeatingMyself - Ignore duplicate provider config entries (Fixes workitem #1800)
EF only allowed one provider config entry for each provider invariant name. However in case of websites it is possible to have a hierarchy of websites which results in squashing all config settings. This results in duplicate provider entries which caused an exception. The fix is to ignore duplicate entries.