NulnessShyness - Altering column from required to optional results in a migration script that does not work on SqlCe
When altering a column from required to optional SqlGenerator for SqlCe and SqlServer did not write 'NULL' depending on the default setting. This worked for SqlServer but not in SqlCe where the nullability of the column being altered would remain required. The fix is to be explicit and always write NULL/NOT NULL when altering columns. We will now do it for both providers.