EF 6.1 - Fix for #1775 - Error in migration (like issue 1773) missing Discriminator Field
Problem: When generating mapping fragments we would sometimes lose discriminator conditions by moving them off of the fragments that ended up in the final output. Additionally, a bug in EntityTypeExtensions meant we would incorrectly generate IsTypeOf fragments for leaf siblings in a hierarchy.
Solution: Rather than move the discriminators, we copy them, so that if the original fragment is output, it has the correct condition. Fixed the issue in IsRootOfSet.

Work Item: 1775