978-1449312961 doesn't have them...but we do! (Almost) (Introduce IndexAttribute and flow into model as IndexAnnotations)
This change introduces the IndexAttribute class together with a convention to read these attributes and generate column annotations based on the the attributes. This includes the ability to merge annotations as necessary when multiple attributes are placed on one property or when multiple properties with attributes are mapped to the same column. Also included is serialization to and from the SSDL XML.

Note that this checkin does NOT cover processing of the annotations by the model differ or the associated generation of indexes by Migrations.