Tweedledum and Tweedledee (765: Allow interceptors to be registered in config file and add interceptor for DbConfiguration.Loaded)
The general change here is to allow any interceptor to be registered in the app.config/web.config file. Also, a new DbConfigurationIntereceptor has been introduced with one method which is the equivalent of the DbConfiguration.Loaded event. When combined these two changes mean that the equivalent of a Loaded event handler can now be registered in the application's config file. Given this, the current mechanism for registered in Loaded event handlers in the config file has been removed.