Insurance for the banana stand... (1937: Update-database should not delete database if any migrations were successfully applied)
The problem in this case was that the database didn't exist before the whole Migrations process started and so the previous fix was returning false for HasMigrations(). This meant that the database was still being deleted. Solution is to set an explicit flag when some statements have been successfully committed and switch delete-or-not behavior based on that flag.