Sirius Cybernetics Corporation Complaints Division strikes again... (Fixes for two issues with TPH column sharing)
Issue 1964: Cannot merge FK properties into single column
The problem here was that the two associations in s-space were still referencing the unmerged columns. However, having an association in s-space here doesn't work anyway because it results in two database constraints which can only be satisfied if the dependent also matches the PK for the other relationship type. Such a match would usually only happen accidentally. The solution is to remove the associations from s-space like we do for similar TPC mappings.

Issue 1924: Shared columns are not made nullable
This means that mappings don't work when there is an abstract base class because a value is required for the column. The fix is to make the merged column nullable. (There may be some cases where this is not necessary, but we chose to take the simple route and always make the column nullable when it is shared.)

Share and Enjoy!
Share and Enjoy!

Share and Enjoy!

Share and Enjoy!

Share and Enjoy!

Share and Enjoy!