Never Mind the Collections, Here's the Nav Pistols (1977: Revert nav prop check back to IEnumerable)
From bug: In EF6 we switched a check that navigation properties are IEnumerable to one that checks if they are ICollection. This is because almost nothing works if the property is IEnumerable but not ICollection and the intention was to help people discover why their code would likely not work. However, as this thread indicates "almost nothing" is not the same as "nothing":  https://entityframework.codeplex.com/discussions/483925#post1150336