High-CH3(CH2)6CH3 - speed improvement re-write of MetadataCollection<T>
Most of our performance investigations during the development of Entity Framework 6 led to concluding that the performance of lookups on MetadataCollection<T> was suboptimal, in many cases composing 10% to 20% of the time spent during the execution of a test. Its implementation also led to functional bugs that were not easy to anticipate.

This re-write of MetadataCollection<T> is based on the following principles:
- Speed-optimize for the main case: lookup by name (case-sensitive).
 - whenever case-insensitive lookups are made, it's okay to not be super-fast.
 - whenever index lookups are made, behave as close to a List<T> as possible.
- Maintain functional equivalence with the previous implementation.

The re-write was done from scratch by copying the type signature of MetadataCollection<T> (public + internal methods) and providing a basic implementation following the principles stated above. In many cases the code of the methods remained the same as it was before the re-write. The resulting class has roughly half the amount of lines of code, is easier to read, easier to maintain, remains functionally compatible and most importantly: is faster.

As an example of the improvement perceived: materializing 31465 SalesOrderHeader entities from AdventureWorks went from 5673 milliseconds to 4602 milliseconds (median value of 10 runs, release build, high-performance power setting, Windows 8.1 64-bit, .NET 4.5.2, Intel Xeon W3530 @ 2.8GHz). This represents almost a 20% improvement.

All tests passing 100%.