No children in bar... (683: EF loads entire child collection when item is added to context)
The fix is to disables lazy loading when accessing navigation properties internally for fixup. Note that this is technically a breaking change since some application may be relying on the behavior that related entities are loaded at this time instead of later when the navigation property or state manager is accessed to get them. However, this seems unlikely and easy to fix so we have decided to take the change.