Morse No. 165, Saskatchewan... (1986: Allow application to switch on logging without recompile)
The ability to register interceptors in the config file makes it possible to turn on logging without recompiled an app. However, there must be a suitable interceptor to register and DatabaseLogFormatter doesn't work because its constructors require a delegate which cannot be specified as a constant expression in the config file.

This is a simple logger that solves this problem by providing constructors that allow logging to the console or a file. It can also be used from code, but its main purpose is config file registration like:

<interceptors>
  <interceptor type="System.Data.Entity.Infrastructure.Interception.DatabaseLogger, EntityFramework"/>
</interceptors>

for logging to the console, or:

<interceptors>
  <interceptor type="System.Data.Entity.Infrastructure.Interception.DatabaseLogger, EntityFramework">
    <parameters>
      <parameter value="C:\Stuff\LogOutput.txt"/>
    </parameters>
  </interceptor>
</interceptors>

for logging to a file.