DDEXBabelFish - throw a better message when translating runtime to design time connection string if DDEX is not installed (Work item: #1838)
Some databases providers (most notably SqlServerCe) have a notion of a runtime and design time connection strings. Connection strings for these databases need to be translated and the designer is using the DDEX provider to do this. However if the DDEX provider is not installed (which is always the case for SqlServerCe in Visual Studio 2013 since Visual Studio 2013 no longer supports SqlServerCe) we would just show a message saying that the translation failed which was not very helpful in understanding why. After the fix if translation fails we check if the DDEX provider is present and if it is not we will add this to the error message.