Partial fix for #1898 "Children could not be evaluated" when enumerating Results View in debugger
As explained in the MSDN documentation at http://msdn.microsoft.com/en-us/library/z4ecfxd9.aspx, when the debugger evaluates an expression in break mode (e.g. a property added to the watch window or the results of the methods printed in the immediate window) all but the current thread of execution are frozen.
Lazy<T>.Value calls Debuger.NotifyOfCrossThreadDependency() which tells the debugger that other threads might need to run temporarily in order to perform the evaluation correctly. If other threads are run temporarily that may have unexpected effects on the state of the program and for that reason the debugger will by default abort evaluation at that point.

The main goal of this change is to remove the use of Lazy<T> in LazyEnumerator<T> to reduce the chances of the debugger aborting evaluation, effectively re-enabling the visualization of query results while debugging (e.g. in the watch or the immediate windows).

Note that with this fix is only partial: if a retrying execution strategy is being used the bug will still repro.

In addition to the fix I have made a few simplifying refactorings around LazyEnumerable<T> usage and the corresponding changes to LazyAsyncEnumerable<T>, as well as a couple of code cleanups.