SkipAPageSkipAPage - handling Runtime Config Page for Code First from DB
CodeFirst from DB is only supported for EF6 so we should never show EF5 option (even disabled) on the page. This boils down to two results - we either want to skip the page if the EF6 provider for selected connection is available or we want to show the error page with the message and EF6 option disabled if the EF6 provider is not available.