If my measurements are correct, this change, together with the new MetadataCollection<T>, improves performance of materialization in the repro project by roughly 35%. It looks kind of "hacky" but there doesn't seem to be a better option, short of re-designing some of the structures/functionality.
TakeSnapshotOfForeignKeys() uses the metadata heavily, going from C space AssociationType space to O space AssociationType and viceversa, and trying to find the association set corresponding to an association type and an end entity set. The last one is actually a linear iteration through all base entity sets in the container. These operations are done for every single entity being materialized. This change optimizes them by storing references to association types and association sets in a way that allows for faster lookups.

RelationshipManager.GetRelationshipType is called to determine the O space association type corresponding to a C space association type. When the method is called for the first time the code iterates through the association types in C space (read-only), assigns them indices and stores the references into the MetadataOptimization._csAssociationTypes array. As the corresponding O space association types are discovered/created they are stored into the similar MetadataOptimization._osAssociationTypes array after the corresponding index (from the corresponding C space association type) is assigned. Subsequent operations that require going from C space to O space and viceversa simply do array access.

Finding an association set corresponding to an association type and an end entity set is optimized similarly, by first building an array of association sets corresponding to the association type with the same index, and then doing array access. An element of MetadataOptimization._csAssociationTypeToSetsMap can be null, an AssociationSet or an array of AssociationSets (for MEST scenarios) hence the type object[].

To be able to integrate this with the existing implementation I had to refactor some of the existing methods in RelationshipManager, RelatedEnd mainly by adding overloads and moving common functionality into helper methods.

It is possible to get some smaller gains if optimizing further (for example having two MetadataOptimization.FindCSpaceAssociationSet methods with similar implementation, instead of one with a Predicate will yield about 2-3% speed increase), but not sure if worth pursuing.
This change optimizes existing functionality without changing the main logic. It may be possible to achieve bigger gains by revisiting the overall design, but not sure how much time that would take.