Too much of a good thing... (2021: Throw better exception when navigation property implements ICollection<T> more than once)
The setup is that the navigation property is declared as a type that implements ICollection twice—once for a concrete type and once for an interface that is implemented by the concrete type. For example:

public interface IRole
{
}

public interface IRoleCollection : ICollection<IRole>
{
}

public class RoleCollection : List<Role>, IRoleCollection
{
}

public class Role : IRole
{
}

We currently don't support this although it might be possible to do so in the future. For now the change is just to throw a better exception message when such a model is used.