Tweedledee and Tweedledum... (2033: IDbDependencyResolver.GetServices not called for IDbInterceptor in latest 6.1 builds)
The problem here was that making Loaded an interceptor resulted in a circular dependency. That is, interceptors need to be requested before Loaded is called so that Loaded interceptors are obtained and can be executed, but if calling Loaded adds new interceptors then these will be missed. The fix is to ask for interceptors again after Loaded has been called and register any new ones at that time.