It's pasta for dinner... (Fixes for 2029 and 2034)
2029: Transaction Commit Failure :: commit failure handler opens additional concurrent connection(s) during initialization

The problem was that Migrations by default always creates new connections to do things. In this case we needed to use the same connection which required a hook to allow the existing connection to be passed in.

2034: Transaction Handler :: unhelpful exception thrown when using Transaction handler and transient failure occurs during operation on TransactionHistory table

The problem was that the transaction object didn't exist because there was a failure creating it. If the transaction object doesn't exist, then no handling can be done for it, so the fix is to check for null before attempting to use the object.