One EF: Honor table names and schemas when a primary key is inferred and the definition is created as read-only
Issue:
When a table does not have a primary key defined, the key is inferred and the definition created as a read-only table. During this process, the table name and schema are removed from the model. This resulted in Code First entities being generated that did not map back to the original view.

Solution:
Before the table name and schema are removed from the model, they are added as annotations. In addition to looking for the table name and schema in the model, we will also look for these annotations. The generated code will then be configured correctly.

Verification:
Manually verified -- the tests use Code First which doesn't have an easy way of adding annotations to store-space entity sets.

Work Item: 2096