This commit provides a fix for issue 633 (https://entityframework.codeplex.com/workitem/633): Lack of UNION/UNION ALL flattening causes errors due to queries "nested too deeply".
The primary difficulty with making the fix came from the fact that, by the time nested union all queries get to SQL gen, each set operator has been wrapped in a "no-op" projection by the CTreeGenerator.
I couldn't find an easy way to detect whether a particular projection was one of these no-ops, so I modified CTreeGenerator to avoid creating the extra projections in the case of UNION ALL and INTERSECT trees.
The change in SqlGenerator is then fairly simple; instead of rendering a set operator as Visit(left), separator, Visit(right), we first expand the tree to get a list of "leaf" nodes and render as Vist(leaf[0]), separator, Visit(leaf[1]), separator, ...
Since UNION ALL has some special-case behavior for Sql 8.0, I disabled the tree expansion logic for that version so that we fall back to something equivalent to the previous code.