One EF: Ignore foreign key column ordering when configuring the model
Issue:
Primary key column order is not specified in the provider model. Instead, we assume that it is the same as the table column order. Foreign key column order is, however. This can result in the primary key and foreign key column order getting out-of-sync.

Solution:
Ultimately, the database order doesn't matter to Entity Framework so long as things are consistent. To keep the orders in sync, table column order will also be used for foreign key columns.

Verification:
Manually verified -- creating and reverse engineering a database is not worth adding to the automated tests.

Work Item: 868