BeingOptimistic - An optimistic fix for workitem #1143
There seems to be a disconnect between DSL and the view model/model and when the view model/model is being disposed DSL may pump messages we are not able to handle because view model no longer exists. In this particular case the fix proposed by David seems safe and should fix this particular crash since we are not using the disposed editing context in any upstream caller. Note that we are curing symptoms and the root cause so if we happen to have other event handlers that are being invoked after the view model/model was disposed that rely on model we may still see crashes.