00ecef70 079c5769 Microsoft_Data_Entity_Design_Model!Microsoft.Data.Entity.Design.Model.EntityDesignArtifact.GetRootNamespace()+0x18
if some corner cases (in this case if the edmx file contents was removed outside VS) the XDocument representing the edmx file could be either empty or null. When EF Designer tried reloading such a file it would throw a NullReferenceException during validation which would make VS crash. The fix is to check for null and treat the document as invalid.