Dexter takes a fall... (2172: FixupReferencesByForeignKeys throws NullReferenceException in EF 6.1.0)
This is a regression caused by a fix that went into 6.1 (Bug 1142) to handle correct fixup when an entity has two FK relationships between the same entities. The underlying problem is that we store a map of FK value to entries that need fixup. But in this case there are actually two identical entries in this map--one for each relationship.

The previous fix handled this by making sure that an entry would not be removed if it was used by another FK. This prevented the entry from being removed when it was still needed.

The new bug is because now the entry may never get removed because on both attempts to remove the other relationship prevents it.

The new solution is to store both the relationship (represented by the RelatedEnd) and the state entry.