ThrowIfYouFeelLikeItOnError: false  (2051: Avoid throwing when assembly version of CLR type changes)
The code that loads CLR types for a model explicitly sets throwOnError to false so that the model load does not fail if a type is missing. This allows Migrations to continue to load a model even after CLR types have changed. However, for some reason if the assembly version is changed then Type.GetType throws a FileLoadException even when throwOnError is false. According to the documentation it can also still throw TargetInvocationException and BadImageFormatException. So the fix is to ignore these exceptions if they are thrown just as we ignore other reasons why the type could not be loaded.

This doesn't prevent the model from having a null CLR type annotation, which for non-Migrations scenarios could result in NullRefException in other parts of the stack. However, this is not different from 6.1 and deciding when and how to throw for a null CLR type is a bigger change that probably doesn't meet the bar right now.