Existential crisis... (Fix issues in database existence checking)
For 2113 The issue is that if a user does not have "view any database" permission then a connection to master will succeed but querying sys.databases will not return any results. We therefore conclude that the database doesn't exist--but it actually does, it just can't be seen when connected to master. The solution is to use the db_id() function which will return non null even when the user does not have "view any database". The catch is that db_id() doesn't work on SQL Azure for anything other than the current database, so if db_is() returns 0 we do the query for sys.databases. This should always work on Azure because Azure doesn't have server class securables, which means it is not possible to deny 'view any database' permissions.

For 2068 the SQL Azure issue is that for SQL Azure a user created just in the target database cannot connection to master when using ADO.NET. (Interestingly SQL Management Studio does allow the same credentials to connect to/query master.) This case was actually handled correctly, except that Exists caused context initialization, which in turn sometimes asks for a provider manifest token, which in turn attempts to connect to the database. If the database exists then this works fine. If the database doesn't exist then an attempt is made to connect to master to get the manifest token. This would then throw, resulting in Exists throwing instead for returning false. The fix for this is to use the provider directly to do existence checks rather than initializing the context and using the ObjectContext method.

For 2068 the open connection issue resulted in an attempt to use a master connection when credentials may have been stripped out. The fix is to short-circuit existence checking if the connection is open since opening the connection implies that the database exists.