CodePlex 2181: Null reference exception in view gen in EF 6.0+ after upgrading from 6.0-beta1
Code in view generation that builds an update view containing a case on the types in a hierarchy assumes that a non-nullable column must have a default value. A change from 2012 made the Discriminator column not nullable (used to be nullable) without specifying a default. The scenario that hits that code path detects that a default value is not specified and attempts to report a view generation error but fails with null reference exception while building the error message.
The fix removes the line of code that appends additional information to the error message, and which causes NullReferenceException, because the message already contains enough information to diagnose the corresponding issue. The fix also specifies a default value of "(Undefined)" for a Discriminator column to satisfy the assumptions in the view generation code.