CodePlex 2160: group by on nullable reference results in empty group
Code in Rule_FilterOverLeftOuterJoin and Rule_FilterOverOuterApply promotes a LeftOuterJoin, OuterApply to InnerJoin and respectively CrossApply if the filter predicate does not preserve nulls for the right table. One of the cases being handled is a simple EQ predicate. However, the equality predicate is potentially expanded in the NullSemantics phase to account for NULLs. Thus the two FilterOpRules need to be deferred until after the NullSemantics phase. We already had code to defer until after the JoinElimination phase to allow that phase to eliminate redundant joins.
The fix updates the mechanism for deferring transformation rules to be a bit more generic, adds a Deferred transformations phase after the NullSemantics phase and updates the code to defer them if conditions are met.