Debuggering up the builder... (Discard ModuleBuilder after it is corrupted by debugger thread aborts)
EntityProxyFactory maintains a thread-safe dictionary of created proxy types. This ensures that if a proxy type has not been created, then only one thread gets a chance to define and create it. As part of that process a ModuleBuilder and its associated stuff is used to do the actual Ref.Emit work. This ModuleBuilder is cached such that only one is used per dynamic proxy assembly. This ModuleBuilder itself has state that is gradually built up as the type is defined. If the whole process succeeds then the thread safe dictionary mentioned above prevents any attempts to define the type again.

The problem comes when some part of the type building process using ModuleBuilder fails. This can happen after the type has been added to the module. In such a situation the cached ModuleBuilder contains a definition for the type—although it is probably not valid because the process of fully building it did not succeed. Since the process aborts the thread-safe dictionary still does not contain the proxy type. This means that the next time the same proxy type is needed the whole builder process starts over again. But since this process is using the cached ModuleBuilder it means that the partially defined type is already in the builder and hence a duplicate type exception is thrown.

In the repro we have the reason that the proxy type build fails is that the debugger does ToString on the query. This causes the query to be translated, which in turn does proxy creation. This is fine, but at some point in the proxy creation the debugger aborts the thread, which then causes the exception observed when the app code runs again. This is why with the current repro code the debugger is required to repro the issue.

It is feasible to imagine that the issue could happen outside of the debugger if something goes wrong in proxy creation. However, this would normally just cause the application to crash. If the exception is caught, then the app may continue, but this is only useful if the reason that proxy creation failed doesn’t occur again the next time a proxy is needed. So it seems unlikely that a running app can deal with bad proxy creation.

The solution is to catch any exception that happens when the type is being built, discard the active ModuleBuilder, and then re-throw the exception. This works since it appears that creating a new ModuleBuilder for the same assembly allows types to be added to an already existing dynamic assembly. It is less risky than always creating a new ModuleBuilder for each type because we still get the caching perf in the normal case and if there is a problem with creating a new ModuleBuilder then it will only manifest in cases where this issue is actually hit.